+++ quickspec/mps/lin.py
@@ -1,3 +1,3 @@
-from camb import mps_lin_camb as camb
+from mps_camb import mps_lin_camb as mps_camb
 from eihu import mps_lin_eihu as eihu
 from bbks import mps_lin_bbks as bbks
